refactor: split src/main.rs into db, models, and handlers modules
The single src/main.rs file had grown to contain the CLI args, the
MMDB reader statics and initialization, the response structs, and every
route handler. Split it into focused modules for readability and to
match the already-present (but empty) src/handlers/ directory:

- src/models.rs: Serialize/Deserialize response structs (Asn, Country,
  RequestedAddress, AsnResponse, CountryResponse) and their constructors.
- src/db.rs: the OnceCell reader statics, init_reader/init_mmdb, and the
  get_country/get_asn lookup helpers.
- src/handlers/{mod,ip,asn,country}.rs: the axum route handlers split by
  resource, reusing the shared db and models modules.
- src/main.rs: trimmed to arg parsing, router wiring, the tokio entry
  point, and the test module which now imports handlers by path.

No behavior changes; the route handlers, lookups, and tests are
unchanged aside from import paths. cargo build succeeds and all 12
tests pass under cargo test --all-features.
